+++ prowler/lib/cli/parser.py
@@ -10,7 +10,6 @@ from prowler.config.config import (
     default_fixer_config_file_path,
     default_output_directory,
     finding_statuses,
-    get_default_mute_file_path,
     valid_severities,
 )
 from prowler.providers.common.arguments import (
@@ -326,14 +325,11 @@ Detailed documentation at https://docs.prowler.com
 
     def __init_mutelist_parser__(self):
         mutelist_subparser = self.common_providers_parser.add_argument_group("Mutelist")
-        provider = sys.argv[1] if len(sys.argv) > 1 else "aws"
         mutelist_subparser.add_argument(
             "--mutelist-file",
             "-w",
             nargs="?",
-            # TODO(PRWLR-3519): this has to be done in the provider class not here
-            default=get_default_mute_file_path(provider),
-            help="Path for mutelist yaml file. See example prowler/config/<provider>_mutelist.yaml for reference and format. For AWS provider, it also accepts AWS DynamoDB Table, Lambda ARNs or S3 URIs, see more in https://docs.prowler.cloud/en/latest/tutorials/mutelist/",
+            help="Path for mutelist YAML file. See example prowler/config/<provider>_mutelist.yaml for reference and format. For AWS provider, it also accepts AWS DynamoDB Table, Lambda ARNs or S3 URIs, see more in https://docs.prowler.cloud/en/latest/tutorials/mutelist/",
         )
 
     def __init_config_parser__(self):
